// packages/evaluatorq/src/generated/openresponses/types/FunctionCall.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { FunctionCallStatus } from "./FunctionCallStatus.js";

export const functionCallTypeEnum = {
    function_call: "function_call"
} as const;

export type FunctionCallTypeEnumKey = (typeof functionCallTypeEnum)[keyof typeof functionCallTypeEnum];

/**
 * @description A function tool call that was generated by the model.
*/
export type FunctionCall = {
    /**
     * @description The type of the item. Always `function_call`.
     * @default "function_call"
     * @type string
    */
    type: FunctionCallTypeEnumKey;
    /**
     * @description The unique ID of the function call item.
     * @type string
    */
    id: string;
    /**
     * @description The unique ID of the function tool call that was generated.
     * @type string
    */
    call_id: string;
    /**
     * @description The name of the function that was called.
     * @type string
    */
    name: string;
    /**
     * @description The arguments JSON string that was generated.
     * @type string
    */
    arguments: string;
    status: (FunctionCallStatus & any);
};