// packages/evaluatorq/src/generated/openresponses/types/FunctionCallOutput.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { FunctionCallOutputStatusEnum } from "./FunctionCallOutputStatusEnum.js";
import type { InputFileContent } from "./InputFileContent.js";
import type { InputImageContent } from "./InputImageContent.js";
import type { InputTextContent } from "./InputTextContent.js";

export const functionCallOutputTypeEnum = {
    function_call_output: "function_call_output"
} as const;

export type FunctionCallOutputTypeEnumKey = (typeof functionCallOutputTypeEnum)[keyof typeof functionCallOutputTypeEnum];

/**
 * @description A function tool call output that was returned by the tool.
*/
export type FunctionCallOutput = {
    /**
     * @description The type of the function tool call output. Always `function_call_output`.
     * @default "function_call_output"
     * @type string
    */
    type: FunctionCallOutputTypeEnumKey;
    /**
     * @description The unique ID of the function tool call output. Populated when this item is returned via API.
     * @type string
    */
    id: string;
    /**
     * @description The unique ID of the function tool call generated by the model.
     * @type string
    */
    call_id: string;
    output: (string | ((InputTextContent & {
        type: "input_text";
    }) | (InputImageContent & {
        type: "input_image";
    }) | (InputFileContent & {
        type: "input_file";
    }))[]);
    status: (FunctionCallOutputStatusEnum & any);
};