// packages/evaluatorq/src/generated/openresponses/types/FunctionTool.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const functionToolTypeEnum = {
    function: "function"
} as const;

export type FunctionToolTypeEnumKey = (typeof functionToolTypeEnum)[keyof typeof functionToolTypeEnum];

/**
 * @description Defines a function in your own code the model can choose to call. Learn more about [function calling](https://platform.openai.com/docs/guides/function-calling).
*/
export type FunctionTool = {
    /**
     * @description The type of the function tool. Always `function`.
     * @default "function"
     * @type string
    */
    type: FunctionToolTypeEnumKey;
    /**
     * @description The name of the function to call.
     * @type string
    */
    name: string;
    description: (string | null);
    parameters: ({
        [key: string]: any;
    } | null);
    strict: (boolean | null);
};