// packages/evaluatorq/src/generated/openresponses/types/ItemField.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { FunctionCall } from "./FunctionCall.js";
import type { FunctionCallOutput } from "./FunctionCallOutput.js";
import type { Message } from "./Message.js";
import type { ReasoningBody } from "./ReasoningBody.js";

/**
 * @description An item representing a message, tool call, tool output, reasoning, or other response element.
*/
export type ItemField = ((Message & {
    type: "message";
}) | (FunctionCall & {
    type: "function_call";
}) | (FunctionCallOutput & {
    type: "function_call_output";
}) | (ReasoningBody & {
    type: "reasoning";
}));