// packages/evaluatorq/src/generated/openresponses/types/Message.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { InputFileContent } from "./InputFileContent.js";
import type { InputImageContent } from "./InputImageContent.js";
import type { InputTextContent } from "./InputTextContent.js";
import type { InputVideoContent } from "./InputVideoContent.js";
import type { MessageRole } from "./MessageRole.js";
import type { MessageStatus } from "./MessageStatus.js";
import type { OutputTextContent } from "./OutputTextContent.js";
import type { ReasoningTextContent } from "./ReasoningTextContent.js";
import type { RefusalContent } from "./RefusalContent.js";
import type { SummaryTextContent } from "./SummaryTextContent.js";
import type { TextContent } from "./TextContent.js";

export const messageTypeEnum = {
    message: "message"
} as const;

export type MessageTypeEnumKey = (typeof messageTypeEnum)[keyof typeof messageTypeEnum];

/**
 * @description A message to or from the model.
*/
export type Message = {
    /**
     * @description The type of the message. Always set to `message`.
     * @default "message"
     * @type string
    */
    type: MessageTypeEnumKey;
    /**
     * @description The unique ID of the message.
     * @type string
    */
    id: string;
    status: (MessageStatus & any);
    role: (MessageRole & any);
    /**
     * @description The content of the message
     * @type array
    */
    content: ((InputTextContent & {
        type: "input_text";
    }) | (OutputTextContent & {
        type: "output_text";
    }) | (TextContent & {
        type: "text";
    }) | (SummaryTextContent & {
        type: "summary_text";
    }) | (ReasoningTextContent & {
        type: "reasoning_text";
    }) | (RefusalContent & {
        type: "refusal";
    }) | (InputImageContent & {
        type: "input_image";
    }) | (InputFileContent & {
        type: "input_file";
    }) | (InputVideoContent & {
        type: "input_video";
    }))[];
};