// packages/evaluatorq/src/generated/openresponses/types/ResponseResource.ts
export const responseResourceObjectEnum = {
    response: "response"
} as const;

export type ResponseResourceObjectEnumKey = (typeof responseResourceObjectEnum)[keyof typeof responseResourceObjectEnum];

/**
 * @description The complete response object that was returned by the Responses API.
 * @example [object Object]
*/
export type ResponseResource = {
    /**
     * @description The unique ID of the response that was created.
     * @type string
    */
    id: string;
    /**
     * @description The object type, which was always `response`.
     * @default "response"
     * @type string
    */
    object: ResponseResourceObjectEnumKey;
    /**
     * @description The Unix timestamp (in seconds) for when the response was created.
     * @type integer
    */
    created_at: number;
    completed_at: (number | null);
    /**
     * @description The status that was set for the response.
     * @type string
    */
    status: string;
    incomplete_details: ((IncompleteDetails & any) | null);
    /**
     * @description The model that generated this response.
     * @type string
    */
    model: string;
    previous_response_id: (string | null);
    instructions: (string | null);
    /**
     * @description The output items that were generated by the model.
     * @type array
    */
    output: ItemField[];
    error: ((Error & any) | null);
    /**
     * @description The tools that were available to the model during response generation.
     * @type array
    */
    tools: Tool[];
    tool_choice: (FunctionToolChoice | ToolChoiceValueEnum | AllowedToolChoice);
    truncation: (TruncationEnum & any);
    /**
     * @description Whether the model was allowed to call multiple tools in parallel.
     * @type boolean
    */
    parallel_tool_calls: boolean;
    text: (TextField & any);
    /**
     * @description The nucleus sampling parameter that was used for this response.
     * @type number
    */
    top_p: number;
    /**
     * @description The presence penalty that was used to penalize new tokens based on whether they appear in the text so far.
     * @type number
    */
    presence_penalty: number;
    /**
     * @description The frequency penalty that was used to penalize new tokens based on their frequency in the text so far.
     * @type number
    */
    frequency_penalty: number;
    /**
     * @description The number of most likely tokens that were returned at each position, along with their log probabilities.
     * @type integer
    */
    top_logprobs: number;
    /**
     * @description The sampling temperature that was used for this response.
     * @type number
    */
    temperature: number;
    reasoning: ((Reasoning & any) | null);
    usage: ((Usage & any) | null);
    max_output_tokens: (number | null);
    max_tool_calls: (number | null);
    /**
     * @description Whether this response was stored so it can be retrieved later.
     * @type boolean
    */
    store: boolean;
    /**
     * @description Whether this request was run in the background.
     * @type boolean
    */
    background: boolean;
    /**
     * @description The service tier that was used for this response.
     * @type string
    */
    service_tier: string;
    /**
     * @description Developer-defined metadata that was associated with the response.
    */
    metadata: any;
    safety_identifier: (string | null);
    prompt_cache_key: (string | null);
};