// packages/evaluatorq/src/generated/openresponses/types/Usage.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { InputTokensDetails } from "./InputTokensDetails.js";
import type { OutputTokensDetails } from "./OutputTokensDetails.js";

/**
 * @description Token usage statistics that were recorded for the response.
*/
export type Usage = {
    /**
     * @description The number of input tokens that were used to generate the response.
     * @type integer
    */
    input_tokens: number;
    /**
     * @description The number of output tokens that were generated by the model.
     * @type integer
    */
    output_tokens: number;
    /**
     * @description The total number of tokens that were used.
     * @type integer
    */
    total_tokens: number;
    input_tokens_details: (InputTokensDetails & any);
    output_tokens_details: (OutputTokensDetails & any);
};